Migrate `pkg init` to `plan init`.
This better reflects the discussed plan in #953.
